+++ internal/bundler/bundler_dce_test.go
@@ -1,0 +1,593 @@
+package bundler
+
+import (
+	"testing"
+
+	"github.com/evanw/esbuild/internal/parser"
+)
+
+func TestPackageJsonSideEffectsFalseKeepNamedImportES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import {foo} from "demo-pkg"
+				console.log(foo)
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				export const foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": false
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/node_modules/demo-pkg/index.js
+const foo = 123;
+console.log("hello");
+
+// /Users/user/project/src/entry.js
+console.log(foo);
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsFalseKeepNamedImportCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import {foo} from "demo-pkg"
+				console.log(foo)
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				exports.foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": false
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/node_modules/demo-pkg/index.js
+var require_index = __commonJS((exports) => {
+  exports.foo = 123;
+  console.log("hello");
+});
+
+// /Users/user/project/src/entry.js
+const demo_pkg = __toModule(require_index());
+console.log(demo_pkg.foo);
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsFalseKeepStarImportES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import * as ns from "demo-pkg"
+				console.log(ns)
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				export const foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": false
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/node_modules/demo-pkg/index.js
+const index_exports = {};
+__export(index_exports, {
+  foo: () => foo
+});
+const foo = 123;
+console.log("hello");
+
+// /Users/user/project/src/entry.js
+console.log(index_exports);
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsFalseKeepStarImportCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import * as ns from "demo-pkg"
+				console.log(ns)
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				exports.foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": false
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/node_modules/demo-pkg/index.js
+var require_index = __commonJS((exports) => {
+  exports.foo = 123;
+  console.log("hello");
+});
+
+// /Users/user/project/src/entry.js
+const ns = __toModule(require_index());
+console.log(ns);
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsTrueKeepES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import "demo-pkg"
+				console.log('unused import')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				export const foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": true
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/node_modules/demo-pkg/index.js
+console.log("hello");
+
+// /Users/user/project/src/entry.js
+console.log("unused import");
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsTrueKeepCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import "demo-pkg"
+				console.log('unused import')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				exports.foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": true
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/node_modules/demo-pkg/index.js
+var require_index = __commonJS((exports) => {
+  exports.foo = 123;
+  console.log("hello");
+});
+
+// /Users/user/project/src/entry.js
+const demo_pkg = __toModule(require_index());
+console.log("unused import");
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsFalseKeepBareImportAndRequireES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import "demo-pkg"
+				require('demo-pkg')
+				console.log('unused import')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				export const foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": false
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/node_modules/demo-pkg/index.js
+var require_index = __commonJS((exports) => {
+  __export(exports, {
+    foo: () => foo
+  });
+  const foo = 123;
+  console.log("hello");
+});
+
+// /Users/user/project/src/entry.js
+require_index();
+console.log("unused import");
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsFalseKeepBareImportAndRequireCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import "demo-pkg"
+				require('demo-pkg')
+				console.log('unused import')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				exports.foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": false
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/node_modules/demo-pkg/index.js
+var require_index = __commonJS((exports) => {
+  exports.foo = 123;
+  console.log("hello");
+});
+
+// /Users/user/project/src/entry.js
+require_index();
+console.log("unused import");
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsFalseRemoveBareImportES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import "demo-pkg"
+				console.log('unused import')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				export const foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": false
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/src/entry.js
+console.log("unused import");
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsFalseRemoveBareImportCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import "demo-pkg"
+				console.log('unused import')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				exports.foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": false
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/src/entry.js
+console.log("unused import");
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsFalseRemoveNamedImportES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import {foo} from "demo-pkg"
+				console.log('unused import')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				export const foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": false
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/src/entry.js
+console.log("unused import");
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsFalseRemoveNamedImportCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import {foo} from "demo-pkg"
+				console.log('unused import')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				exports.foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": false
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/src/entry.js
+console.log("unused import");
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsFalseRemoveStarImportES6(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import * as ns from "demo-pkg"
+				console.log('unused import')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				export const foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": false
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/src/entry.js
+console.log("unused import");
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsFalseRemoveStarImportCommonJS(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import * as ns from "demo-pkg"
+				console.log('unused import')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				exports.foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": false
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/src/entry.js
+console.log("unused import");
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsArrayRemove(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import {foo} from "demo-pkg"
+				console.log('unused import')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				export const foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": []
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/src/entry.js
+console.log("unused import");
+`,
+		},
+	})
+}
+
+func TestPackageJsonSideEffectsArrayKeep(t *testing.T) {
+	expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/entry.js": `
+				import {foo} from "demo-pkg"
+				console.log('unused import')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/index.js": `
+				export const foo = 123
+				console.log('hello')
+			`,
+			"/Users/user/project/node_modules/demo-pkg/package.json": `
+				{
+					"sideEffects": ["./index.js"]
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/entry.js"},
+		parseOptions: parser.ParseOptions{
+			IsBundling: true,
+		},
+		bundleOptions: BundleOptions{
+			IsBundling:    true,
+			AbsOutputFile: "/out.js",
+		},
+		expected: map[string]string{
+			"/out.js": `// /Users/user/project/node_modules/demo-pkg/index.js
+console.log("hello");
+
+// /Users/user/project/src/entry.js
+console.log("unused import");
+`,
+		},
+	})
+}